Made minimal changes to control.py
I assume I am thinking of this right, but I think the data is just being fed into this via the data, and then I think all I needed to do was update the angle. It looks like the heavy lifting was done outside of this function, but I am not 100% sure. I think I just need to say where the gap is